libraries/imagenet/pyproject.toml
[tool.poetry]
name = "imagenet"
version = "0.1.0"
description = "TensorFlow MobileNetV2 demo using imagenet pre-trained network"
authors = ["morten.aslesen <mortendaehli@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.10"
tensorflow = "^2.7.0"
keras = "^2.7.0"
Pillow = "^8.4.0"
numpy = "^1.21.4"

[tool.poetry.dev-dependencies]
pytest = "^6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
